requirements.txt
PyJWT==2.8.0
graphene-django==3.1.3
graphene-file-upload==1.3.0